Changed mouse to set/get int instead of float when it comes to positions

+++ BIS/BIS/System/System.cpp
@@ -214,7 +214,7 @@ void System::mAddEvent(Room * room)
 	EventData event_data = { 0 };
 	for (int i = 0; (i < (int)events_in_room.size()) && (i < 4); i++)
 	{
-		event_data.slots[i] = events_in_room[i]->GetType() + 1;
+		event_data.slots[i] = (float)(events_in_room[i]->GetType() + 1);
 	}
 
 	top_picked_deck->SetEvent(
@@ -246,7 +246,7 @@ void System::mRemoveEvent(Room * room)
 	EventData event_data = { 0 };
 	for (int i = 0; (i < (int)events_in_room.size()) && (i < 4); i++)
 	{
-		event_data.slots[i] = events_in_room[i]->GetType() + 1;
+		event_data.slots[i] = (float)(events_in_room[i]->GetType() + 1);
 	}
 
 	top_picked_deck->SetEvent(
@@ -284,7 +284,7 @@ void System::mUpdateEvents(Room * room)
 	EventData event_data = { 0 };
 	for (int i = 0; (i < (int)events_in_room.size()) && (i < 4); i++)
 	{
-		event_data.slots[i] = events_in_room[i]->GetType() + 1;
+		event_data.slots[i] = (float)(events_in_room[i]->GetType() + 1);
 	}
 
 	top_picked_deck->SetEvent(
